refactor: Move Vue logic
