+++ src/plugin.js
@@ -1,16 +1,20 @@
 import { configureServer, getRenderContext } from './dev/middleware.js'
 
+const name = 'vitedge'
+
 export default () => {
   return {
-    name: 'vitedge',
+    name,
     configureServer, // Provide API/Props during development
     configResolved: (config) => {
-      // Vite-beta>=69 wraps 'alias' in 'resolve'
+      let lib = '/vue' // default
+
+      const file = config.build.ssr ? '/entry-server' : '/entry-client'
+
+      // config.alias is pre-beta.69
       ;(config.resolve.alias || config.alias).push({
         find: /^vitedge$/,
-        replacement: config.build.ssr
-          ? 'vitedge/entry-server'
-          : 'vitedge/entry-client',
+        replacement: name + lib + file,
       })
     },
     viteSsr: {
